move releases under platform interface and simplify calling
