+++ src/platforms/gitlab/release_test.go
